directory-async: set NULL the count progress on cancel
When a file gets moved, nautilus directory cancel the loading of
the attributes of that file, in case that file is in the work queue
of that directory, since it will move to another directory.

For that, we cancel the cancellable associated to the async request
of the attributes of that file.

However, since this is threaded, if some client of the directory kick of
I/O with the nautilus_directory_async_state_changed and the cancellable
didn't reach the callback, the file attributes are still going on, and
we reach an assert when trying to stop them again, since the file
requesting for those attributes is no longer the one we want to stop
loading.

This was causing problems when the race kicks in, for example, the
move to context menu action.

To fix it, apart of cancel the cancellable associated, set the directory
private data files as null, to avoid trying to stop the already stopped
request.

https://bugzilla.gnome.org/show_bug.cgi?id=756253