bookmarks: remove bookmarks manager
Current bookmark manager is mostly a "show the internals in the UI",
with just showing the properties in a text base way.

Users can already accomplish the key functions of the manager through
the sidebar: you can rename, remove, and reorder through drag and drop.

The only thing you can't do with the sidebar is edit the path of a
bookmark, but I would argue that that's problematic anyway (since you
can break bookmarks that way, and it is much less error-prone to
navigate to the location and bookmark it than hand edit a location
field).

The fact that the only way to open the bookmark manager is through a
keyboard shortcut is probably a sign that it isn't really getting used
anyway, and that it isn't worth the trouble of fixing it up.

So remove it and instead, if needed, focus on improving bookmarks
management in the sidebar.

https://bugzilla.gnome.org/show_bug.cgi?id=759504