application: move actions to the same file
We had two files for application related functions, one for actions and
one for the application itself.
Move the actions to the same file so it's easier to scan and share code.
We did something similar to nautilus window for 3.16, and I had this
planned for some time, but didn't get to it before.

Also remove tabs sneakely...