window-slot: sync search visibility when using history
We can return to an old location which happens to be a search
location.
The issue is, we were not updating all the associated properties
of the search when that happened so the window slot and the
files view was in a inconsistent state.

For that, we need the files view to also accept to change to search
locations, in which case it needs to set it's query appropiately
to the query the search location was using.

Thanks to Georges for the initial work on this.

https://bugzilla.gnome.org/show_bug.cgi?id=756183