files-view: add optional menu item for creating links
The menu item for creating links was removed in previous versions of nautilus
since it exposes a concept of the file system that is not really clear.
However, we don't have a working solution yet for the use cases that creating
links is a workaround, so we didn't remove the functionality altogether.

We were allowing link creation with a shortcut and with the middle button while
performing a drag and drop operation. However, some users would need to use a
context menu action instead of a drag and drop operation, which usually is less
convenient and prone to errors.

Since this is demanded, implement the menu action for it and add a gsetting
preference to show it in the context menu for those users who like to have it
there.

Also the new implementation uses the code that is already used for other
operations, improving the implementation compared to the previous one.

In an upcoming patch we add the UI for the preference dialog.

https://bugzilla.gnome.org/show_bug.cgi?id=745575