files-view: add create link menu item for selected files
The menu item for this feature was removed in previous versions of nautilus.
A context menu item for creating links from copied files was added, but some
users prefered to create links from selected files.

Since this is demanded, implement the menu action for it and use the gsetting
added in the previous commits.

https://bugzilla.gnome.org/show_bug.cgi?id=745575