file-operations: always pop undo manager flag
Nautilus undo manager reuses undo information based on a flag that is pushed
when an undo operation begins and popped during the operation itself. In the
case of trashing / deleting files, the flag was not popped in some situations,
which caused the next operations to not be undoable.

Flip operands of a logical operator so the flag is always popped, avoiding
issues caused by lazy evaluation.

A better solution would be to always update the flag internally in the manager,
and only peek its value externally.

https://bugzilla.gnome.org/show_bug.cgi?id=759850