files-view: Fix double-free of file list in scripts menu
When building the scripts menu, update_directory_in_scripts_menu()
gets the list of files in the script directory, filters out hidden
files and sorts the filtered list by display name. Unlike the filtering
step, which returns a new list with ref'ed files, sorting may change
the start of the list, but not the list's actual content. As a result,
the nautilus_file_list_free() call added in commit 864c815479a25 tries
to free already freed memory (i.e. disposed files), resulting in a
crash when any scripts are found.

https://bugzilla.gnome.org/show_bug.cgi?id=760338