files-view: use finalize instead of dispose
Dispose is intended to let the object still valid for external
calls.
Freeing the action group on dispose makes the view somehow invalid,
so some calls were made at that time and were making the view to crash.

This change to dispose was made in commit 347369d18eb6 to fix a crash as
well. However, an upcoming patch fixed it as well as noted in the
commit message.
So although dispose seemed good, it's not if it makes the object
invalid.

On the other hand, instead of dispose we should use destroy, since we
are already using it, and implementing dispose was kind of an error.

In upcoming patches we will move some disconnect_signal calls from
finalized to destroy.

https://bugzilla.gnome.org/show_bug.cgi?id=759717