css: load css that nautilus always need
There is some css that nautilus wants to always use regarding
of the theme that is used.

An example of this is the recently css added for making the desktop
backgrund transparent.
Additionally, the "round button" css is also always wanted.

For that, create a new css file that will have the css we want
to always load and load it in the application initialization.

https://bugzilla.gnome.org/show_bug.cgi?id=760790