search-directory: don't mark as all_files_seen on hits ready
We were marking as all_files_seen when hits were added, so the view
will react on it and display the files that were recently added.

However, that's not consistent, and the view calls done_loading
if all files are seen. This has, as a UI effect, removing the floating
bar, and therefore any hint that the view is still loading/searching.

So let's be consistent here and in the upcoming patch we will fix
this issue with the search.