files-view: display search files while searching
The view doesn't display files while it's loading, and that's
mostly fine and what we want to avoid jumping while seeing new
files.

However, the when searching, it's expected to not all files be there,
and the first ones reported by the search system have a higher score
to appear, so that means it's fine if the user doesn't see files with
low score while files with high score are presented.
On the other hand, is not fine to wait until the search system
stops to show all the files that were found.

So for search we want to show files while they are being loaded, that
means the directory doesn't have the all_files_seen flag set.

Previously, we were solving this issue setting the all_files_seen
flag in the search directory so the view will show them when hits
are added. However, that also means the view thinks the loading is
complete, removing the floating bar and any hint that it was loading.

So what we actually want is special case the search, and show the files
while they are being found.

The previous patch reverted the change to mark the search directory
as all_files_seen to be consistent. This patch makes the view display
files while they are being found.