search-directory: don't reload search
I know that it looks like it makes sense, but don't do it :)

Basically, all the logic for when a the query changes and to
load a new search directory (and therefore reload) is done outside
of the engines or the query. So basically don't touch it here.

What I believe is that reload should start the search again, not only
stop it and invalidate the file attributes of the search directory,
because as you could see it actually stops the search but doesn't
restart it.

I think this change of logic that makes reload not work properly on all
cases was introduced by me when I introduced the concept of restarting
the search engine internally, not only start and stop.

I will take a deeper look in future.