files-view: reduce floating bar delay
I was confused even for a few miliseconds plus the idle time
why there was no searching going on.
So decrease the floating bar delay to the point it only won't appear
if the user wouldn't have time to look for a hint to know if the view
was searching or not.