query-editor: use is_* instead of !is_
An element within the group that complies a property A doesn't mean
the element automatically complies property B if we don't know
previously that those properties are mutual exclusive.

This can bring misbehavior if at some point a property C is introduced
making the other two non mutual exclusive.

Basically, it's better to check for the property you are going to assume
on the code than the opposite mutual exclusive property.
Also, is usually easier to understand what an element is than what is
not.