files-view: don't remove floating bar in an idle
If the view is still loading.

This happens when we schedule an idle to update the selection info
to display it in the floating bar, but selection becomes null.
Then the idle callback will remove the floating bar because the status
is none.

But that is wrong, because the floating bar is used for more than to
display the selection. We use it for displaying whether the view is
loading or searching, so if that happens we were removing the floating
bar.

To fix it, check in the idle callback for the other use case of the
floating bar, that means if it's still loading, and do nothing in that
case since that is taken care in a different code path.