search: use range of dates
This was not implemented. We want to differenciate between a range of
days we want to search for and a specific date, as set on the mockups.

The calendar selector will be specific dates, whether the days selector
will be a "since".

Also on the way fix various things and code preferences.