search-engine-tracker: fix date bounds
With the previous patch we set the correct bound for the dates,
however due to https://bugzilla.gnome.org/show_bug.cgi?id=761455
bug in tracker, we cannot use dates vars defined in the SELECT clause
in the WHERE clause.

After discussing the nautilsu query to tracker, we came up that it's
not needed to use OPTIONAl and coalesce for it, since contentAccessed
is not something that will be set for files, instead the file ontology
should be used.

So just remove the coalesce and move the variables definitions to the
WHERE clause to make them with the correct type.