preferences: add remote and local recursive search
Instead of using a switch in the search popover.

The search popover is meant to be as a temporary filter. That means
that the "Search subfolders" switch that was present there was reset
every time a new search was performed.

Even if the nature of the popover is temporary and therefore should be
understandable that the switch is also temporary, this can bring
confusion in such a sensible matter.

To avoid confusion, add two preferences, one for remote file systems
and one for local file systems to allow the choice to make a recursive
or non recursive search, and remove the switch to avoid frustration.

Also, I expect this choice to be more a permanent one than a temporary
one, as in, I expect users to what they really want is to make a
permanent choice whether they want recursive search or not.

For local file systems, on what I can gather, either wants to emulate
the type-ahead search, because it's file system is slow to perform
a recursive search and will always be, therefore a permanent choice,
or the opposite where the file system of the user is fast enough to
perform a recursive search, which will most of the cases be like that,
and therefore also a permanent choice.

For remote file systems is similar. Either the internet connection of
the user is fast enough for the whole session or use, therefore wants
recursive search always enabled, or it's not, and therefore it doesn't
want recursive search enabled.