file-operations: replace io_scheduler with GTask API
Nautilus file operations are implemented as asynchronous jobs scheduled using
g_io_scheduler. Since g_io_scheduler has been deprecated, these operations
should be using the simpler GTask API. The helper functions used in the
operations have been changed in a previous patch so it is now possible to port
the jobs themselves to the new API.

The job structures are now data for tasks, which are handled by the existing
functions in separate threads. For finalizing the operations, the existing
"job_done" functions are now used as callbacks.

https://bugzilla.gnome.org/show_bug.cgi?id=761549