file-undo-operations: change trashed files matching condition
In Nautilus, undoing trash operations relies on the original path of deleted
files and their deletion time. This time value is obtained after the operation
is performed, which can lead to a mismatch. This mismatch would cause deleted
files to not be restored by an undo operation, as they would not be recognised
as being the originals.

Compare the deletion time from the file in trash with the value stored in the
undo information. If the difference is less or equal than an epsilon value,
consider the files as matching.

https://bugzilla.gnome.org/show_bug.cgi?id=761819