search: use a separate view setting for search
So we can select what type of view do we want for search independently
of what we normally use.

This is needed since we default to switch to list view for search, but
we would like to allow users to select a different view. However,
instead of adding a preference in the preference dialog, we can do it
more straightforward and change the setting when we are in search.

On the way, rework all the enums and views id for a saner code...