file-undo-operations: modularize restoring files from trash
In Nautilus, restoring files from the trash is done based on the trashed files
kept in the undo information. In order to find the matching files that need to
be restored, the contents of the trash are ran against the files in the undo
info. Enumeration of the trash and matching are currently handled by the same
function, which does not allow for any of these operations to be used
separately.

Add a function for synchronously enumerating the contents of the trash. Add a
function for matching files in the undo info with the list of files in the
trash.

https://bugzilla.gnome.org/show_bug.cgi?id=762126