file-operations: check that trashed files are actually in the trash
In Nautilus, trash operations are performed and finalized before the trash is
updated. Because of this, the option to undo trashing becomes available even if
the trash is not aware of new files. If an undo is then requested, the operation
would do nothing, because the files would appear to not exist in the trash. A
better solution would be to wait for the trash to notified, and then check that
the trashed files are actually there. Only then should a trash operation be
considered finalized.

Add a signal to the trash-monitor that is emitted when the trash changes.
Connect trash operations to this signal. On signal emission, asynchronously
check that the files are in the trash. Finalize the job if the check was not
cancelled and successful.

https://bugzilla.gnome.org/show_bug.cgi?id=762126