css: draw a background for the notebook
With recent gtk+ changes, the notebook style and drawing was
changed somehow.

Nautilus had a white background by "luck", and the canvas view
needs to apply it's own background style in order to have a background
at all.

This commit follows what we were doing with the desktop window where we
apply a style class to have a transparent background, but in this case
a white one to fix this issue.

https://bugzilla.gnome.org/show_bug.cgi?id=761965