preferences: reword link and delete actions
We want to make them affect the action itself, but that can confuse
users that were already using them, and probably doesn't worth the
effort.

So for now make clear in the preferences window that what it affects
is showing or not the action in the menu, not the availability action
itself.