window: add notifications for sidebar operations
Recently we removed gtk+ notifications from mount operations in the
gtk+ sidebar in order to move the handling of those to Nautilus.

Now that that is done, implement notifications handling for unmount
sidebar operations. The mount operations notifications are still handled
by gnome-shell, we will see if that continues making sense or not.