pathbar: remove gtk_style_context_invalidate
In Nautilus, a size allocation of the pathbar could cause style context
invalidation for its children. Previously the invalidation had to be done
manually. Since it is now done automatically, checking if it is needed and
explicitly doing it are no longer necessary.

Remove function responsible for invalidating style contexts. Remove the logic
for checking if invalidation is needed.

https://bugzilla.gnome.org/show_bug.cgi?id=762246