window: use slot location instead of loading
We were changing the location of the sidebar every time the slot started
loading.
However, that's dangerous since the slot can start loading few times
even with the old location.

Instead of using the loading property, use the location property and
change the sidebar only when the location actually changes on the slot.

This fixes the sidebar jumping rows with gtk+ 3.20.

https://bugzilla.gnome.org/show_bug.cgi?id=762137