placesview: avoid symbol conflicts
Copy pasting the code from gtk+ has the downside that symbol conflicts
occur when the file chooser is shown, for example, when performing a
move to action.

Only way to dealing with it is either make this public on gtk+ or
renaming the types. We don't want to make it public yet, so the only
option for now is renaming.

Is a tedious task and I hope it won't be needed in the future.
For now, rename the types manually.

https://bugzilla.gnome.org/show_bug.cgi?id=760307