window: use open flags for other locations view
We weren't checking the open flags from the sidebar when the other
locations was clicked.

This was due to the sidebar not adding those flags to the signal.

In a recent commit gtk+ added the missing flags, so now we can check for
those in nautilus and open in the current view, in a tab or in a window
accordingly.

https://bugzilla.gnome.org/show_bug.cgi?id=754743