window-slot: avoid random switching of the view
We were not initializing the view mode before search to the invalid
id, which then was initialized as 0 that it's actually a view id.

This was confusing the checks and making random changes of the view
on the first time the location was changed.

This was introduced as a regression when we introduced the separated
search view mode, which also reworked all of the view id handling.