Search popover: limit options to max 5 years
The list displayed too many options for older dates.

In order to limit this, the max_days variable was decreased to
the number of days corresponding to 5 years

The value of the step for months was updated to the proper value, 90.

The days variable contained an unnecessary offset. To prevent it,
for the first occurrence of each timeslice (month, week, year) there
was added an if clause to remove the offset

Due to limiting the max number of years displayed, there was removed
(a now redundant) else from the function get_text_for_date_range.

https://bugzilla.gnome.org/show_bug.cgi?id=762240