nautilus-file-operations.c: new files from template no longer begin with "Untitled"
When a new document is created using a template, the word "Untitled"
is displayed in front of the template name. For example,
creating a new document with a template called "document.doc"
will create new file named "Untitled document.doc".

This is obviusly a non-practical use of templates and this patch
aims to fix this miss-behavior.

The fix is simple, the word "Untitled" has been removed from the
code when a user creates a new document from a template.

https://bugzilla.gnome.org/show_bug.cgi?id=762774