files-view: hide hidden files when renamed
We were not renaming the file at all when the user changed the
file name to a dot file i.e. .whatever.

This was confusing. A better solution would be to show the new
file name until the user changes the directory, but I'm not sure
how much better is it versus hidding the file straig away, since
it can be even more confusing trying to hide your file and nautilus
not hiding it straig away.

For now hide instantaneously a file when renamed to a hidden file.

https://bugzilla.gnome.org/show_bug.cgi?id=342819