Don't show thumbnails for small zoom levels
When the zoom level is too small, thumbnails become useless, as they are
so small that they don't give any information.

The solution to this problem is to use mimetype icons when the zoom
level is too small

In order to achieve this, there was added a function that returns
false if the zoom is too small, nautilus_thumbnail_is_limited_by_zoom.
Therefore, if it returns false, the mimetype icon will be used.

https://bugzilla.gnome.org/show_bug.cgi?id=752805