files-view: select first on loading only if no previous selection
Avoid to select the first file if there is a selection already.
Although we were checking for the pending selection, it could be that
there is a selection already because the user selected a file.

This will be needed also for an upcoming patch, where we always select
the first file on displaying new files, even if the directory is not
fully loaded.

https://bugzilla.gnome.org/show_bug.cgi?id=762595