files-view: select first file even when directory not fully loaded
Currently we were waiting until the view was fully loaded to select the
first file.

This is fine when only navigating.

However, for search, we actually want to select the first file in order
to allow the user to open right away the most relevant file.

Let's see if this brings problems on icons jumping in large directories,
but I hope it doesn't in the most common cases.

https://bugzilla.gnome.org/show_bug.cgi?id=762595