notebook: avoid updating if no children
We were having a critical when a tab was searching and we closed the
window due to the view reporting it's not loading anymore when the
notebook already removed the slot.
This is due to the slot working independently to the notebook, and the
view reporting to stop loading on dispose, which is fine since the view
needs to report its new status to the window as well. However the window
also update the notebook when is the notebook that removes the slot
and therefore dispose the view.

To avoid that, check on the window if the notebook still contains
the slot or not, and update only in that case.