shell-search-provider: make it work again
We removed the search action with a rework on the whole handling
of views/slots/search-query handling to decouple the code better
and not expose API that is not needed outside.

The problem is that we actually needed a way to search from the
application, since gnome-shell search provider communicates in that way.

However we missed this since it was just an action in the application,
which made us don't catch this.

Now we allow a search in the whole stack but in a cleaner and direct
way to not be in the same situation in the future.

This patch use that to make the shell search provider work again.

https://bugzilla.gnome.org/show_bug.cgi?id=762076