window-slot: workaround search bar double border
We currently have a double border in the view due to the search bar
using a revealer, which always allocates 1 px and therefore shows the
bottom border.

Since to fix it we would need widgets able to allocate 0 px, which is
current impossible, we workaround it with a CSS style that moves the
content 1 px upward.