files-view: added alt+return shortcut for properties
With nautilus 3.16 the properties shortcut was changed from
alt+return to ctrl+i.

alt+return is a known shortcut in other applications for showing
file properties, so it was proposed to add alt+enter and altgr+enter
as shortcuts in bug 662081.

nautilus_application_add_accelerator() does not accept multiple shortcuts
per action, so I used gtk_application_set_accels_for_action() and bound
view.properties to the array open_properties.
Also the additional shortcut was added to the shortcuts overview.
Unfortunately, I couldn't find a reliable way to bind altgr for the
shortcut, as modifiers are dependend on the X server mapping.

https://bugzilla.gnome.org/show_bug.cgi?id=662081