application: handle notifications internally
In Nautilus, operations such as unmounts create notifications that are not
withdrawn on application closing. This leads to issues such as notifications
that persist even after reboot. In order to fix this, the application should
keep track of the notifications sent in order to withdraw them when it is
closed.

Add helper functions for sending and withdrawing notifications. Add "shutdown"
signal handler for removing notifications on application closing.

https://bugzilla.gnome.org/show_bug.cgi?id=763129