window, progress-persistence-handler: change notifications API
In Nautilus, notifications are sent using GApplication functions and are not
tracked. This leads to notifications that are not being cleared when Nautilus is
closed. In order to fix this, sending and withdrawing notifications should be
done with the methods of NautilusApplication. These methods automatically track
and clear them when the application is shutdown.

Replace GApplication functions for sending and withdrawing notifications with
corresponding NautilusApplication functions.

https://bugzilla.gnome.org/show_bug.cgi?id=763129