window: hide delete notification on undo
In Nautilus, after a delete operation, an undo notification is shown to the user.
If the operation is undone by other means (such as the Ctrl+Z shortcut), the
notification is not automatically hidden, which can lead to an invalid undo
operation. In order to fix this, hide the notification on an undo operation.

Hide notification in the "undo-changed" signal handler.

https://bugzilla.gnome.org/show_bug.cgi?id=761546