file-undo-manager: fix trash undo invalidation
In Nautilus, when the Trash is emptied, trash undo / redo operations are
invalidated by the undo manager. However, only the undo operations should be
invalidated, because a redo is not affected by the state of the Trash. In order
to fix this, change the invalidation condition to only target undo.

https://bugzilla.gnome.org/show_bug.cgi?id=763134