files-view: add Open With App menu item for all items
The menu item "Open With Application" is not available for most files.
This is a valid menu item for files and folders of all types where the
selected item can be opened in some other application.

To fix it : The condition in the "src/nautilus-files-view.c" that adds
open with application action has been modified so that this menu item
appears in the context menu for all cases (except when the right click is
triggered in empty space)

https://bugzilla.gnome.org/show_bug.cgi?id=566477