content-bar: don't auto destroy the widget
We were auto destroying the content bar if the mime type couldn't
be handled.

This is wrong since the caller is not aware of anything, trying to show
or do other things with the destroyed widget.

I'm not sure how this even work, if it worked at all.
Previously it wasn't crashing because gtk+ was failing gratefully,
however, with the gadget rework, this makes gtk+ crash.

We fixed this in the previous patch, only creating the content bar if
we can handle the myme types.
This patch adds a few more guards and, in any case, avoid to destroy the
widget inside the widget itself.

https://bugzilla.gnome.org/show_bug.cgi?id=762703