placesview: auto generate code
We have been manually copying the code inside nautilus since we
introduced the places view.

It has been a pain to maintain, mostly because we needed to remove the
bits that only work inside gtk+ and instead either remove them or make
a substitution. But that was doable.

However, it reached a new level when we realized that we use the file
chooser inside nautilus, for the "move to" and "copy to" actions, which
create makes symbols clash. So we needed to rename all the symbols in
those files.

Instead of making it manually, create a shell script that fetches gtk+
repository and make the appropriate substitutions, deletions and what
not.

https://bugzilla.gnome.org/show_bug.cgi?id=760307