pathbar: always show the sliders
The style cannot be done in a way that makes the path bar with a
permanent visual glitch.
Follow what the file chooser does and always show the sliders to prevent
that, although we are reclutant to any solution, none of them are good.

Let's make the new path bar for 3.22.