desktop: move to a different binary
We wanted to do this for long time. This will allow to handle the
desktop process in a different binary.
The ultimate goal is to make the desktop code completely split from
nautilus code.

This is the first and minimal step towards that goal.

In this patch we create a desktop application separated from nautilus
application, and remove the desktop handling in nautilus application.

https://bugzilla.gnome.org/show_bug.cgi?id=712620