file-operations: fix formatting of delete progress
A simple operator precedence mishap prevented the deletion rate during
long deletion operations from being displayed properly. Also, the
formatting of the message is subtly different from the others of this
kind because we don't have a predefined formatting token for a file
count, and it ended up missing a space character.

https://bugzilla.gnome.org/show_bug.cgi?id=759986