window-slot: use inheritance for other locations view
We need to special case the other locations view when using that
location, since it's not a files-view and doesn't support several things
that we usually support, like the changes between icon view and list
view.

Also we specifically special case its creation in window slot and we
disable few actions that are not available on it.

This patch creates a other locations slot, which will handle all of it.
The class that is responsible of creating one type of slot or another is
the window, and will use a vfunc that will request whether the slot
handles a location or not and will act accordingly.

In upcoming patches we will move all the special casing of this and the
desktop in the window slot to its respective subclasses now that we have
everything ready.

https://bugzilla.gnome.org/show_bug.cgi?id=712620