files-view: remove the floating bar
The floating bar was a bad choice since it's inception. Not
only it hovers content, it's behavior is inconsistent and the
usability of it is very poor.

After the introduction of the actionbar, most of the data that
we displayed through the floating bar was moved to the actionbar,
making the floating bar obsolete.

This commit follows up and completely removes the floating bar.

https://bugzilla.gnome.org/show_bug.cgi?id=767874