files-view: don't hold a reference to a view for scripts
I'm not sure the scripts should hold a reference to the view itself...
Just remove it as we did with the templates in the past.

What we can do in future is add a weak reference, but I'm still unsure
this is necessary here.

For now fix this since it causes crashes when using scripts.