window-slot: fix shortcuts for changing view mode
In Nautilus, the action for changing the view mode takes an unsigned integer as
parameter in order to determine to which mode to switch. The parameter's type
was not specified in the action name, which lead to the shortcuts not working.
In order to fix this, add the missing type.

https://bugzilla.gnome.org/show_bug.cgi?id=764167