do not switch focus to new tabs
In Nautilus, when a location is opened in a new tab, the new tab gets
automatically focused. This behavior is not consistent with other tab-based
applications, such as browsers - Epiphany, Firefox, Chromium to name a few. It
is also a regression, since this behavior was not present in 3.14. In order to
fix this, set a flag so the new tab is not made active.

https://bugzilla.gnome.org/show_bug.cgi?id=764035